+++ internal/source/file/file.go
@@ -6,6 +6,8 @@ import (
 	"os"
 	"path/filepath"
 	"strings"
+
+	"primamateria.systems/materia/pkg/source"
 )
 
 type FileSource struct {
@@ -32,9 +34,14 @@ func NewFileSource(c *Config) (*FileSource, error) {
 	}, nil
 }
 
-func (f *FileSource) Sync(ctx context.Context) error {
+func (f *FileSource) Sync(ctx context.Context, opts source.SyncOpts) error {
 	if _, err := os.Stat(f.Destination); os.IsNotExist(err) {
 		return fmt.Errorf("source destination path %v does not exist", f.Destination)
+	}
+	if opts.Subpath != "" {
+		if _, err := os.Stat(filepath.Join(f.RemoteRepository, opts.Subpath)); os.IsNotExist(err) {
+			return fmt.Errorf("source subpath %v/%v does not exist", f.RemoteRepository, opts.Subpath)
+		}
 	}
 	entries, err := os.ReadDir(f.Destination)
 	if err != nil {
@@ -45,8 +52,9 @@ func (f *FileSource) Sync(ctx context.Context) error {
 			return fmt.Errorf("error syncing filesystem: can't clear path: %w", err)
 		}
 	}
+	source := filepath.Join(f.RemoteRepository, opts.Subpath)
 
-	repoFS := os.DirFS(f.RemoteRepository)
+	repoFS := os.DirFS(source)
 	err = os.CopyFS(f.Destination, repoFS)
 	if err != nil {
 		return fmt.Errorf("error syncing filesystem: can't copy fs: %w", err)