+++ pkg/source/main.go
@@ -6,7 +6,7 @@ import (
 )
 
 type Source interface {
-	Sync(context.Context) error
+	Sync(context.Context, SyncOpts) error
 	Close(context.Context) error
 	Clean() error
 }
@@ -14,6 +14,11 @@ type Source interface {
 type SourceConfig struct {
 	URL  string `toml:"url" json:"url" yaml:"url"`
 	Kind string `toml:"kind" json:"kind" yaml:"kind"`
+}
+
+type SyncOpts struct {
+	Revision string
+	Subpath  string
 }
 
 func (c SourceConfig) String() string {