+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Best effort list of major changes and bugfixes
 - feat: simplified volume compression. `containers.compression_command` and `containers.compression_suffix` have been simplified into one `containers.compression` setting that can be blank, `gzip`, or `zstd`
 - refactor/note: volume dump file are now in the format `volumename-volume.tar(.gz/zst)`.
 - feat: `PreCommand` and `PostCommand` component manifest settings for running scripts before or after an update. Experimental.
+- feat: Remote components can specify a `subpath` to use.
 
 ## 0.6.4
 - No user facing changes